intent-python/requirements.txt
aniso8601==9.0.1
APScheduler==3.9.1.post1
asgiref==3.5.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bert-extractive-summarizer==0.10.1
blis==0.7.5
breadability==0.1.20
build==0.7.0
cached-property==1.5.2
catalogue==2.0.8
certifi==2021.10.8
cffi==1.15.0
chardet==5.1.0
charset-normalizer==2.0.7
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.4
contourpy==1.0.6
cryptography==36.0.2
cycler==0.11.0
cymem==2.0.6
Cython==0.29.28
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
docopt==0.6.2
docx2txt==0.8
en-core-web-md==3.0.0
en-core-web-sm @ file:///C:/Users/sandeep.harith/Downloads/en_core_web_sm-3.0.0.tar.gz
et-xmlfile==1.1.0
exchangelib==4.5.1
executing==1.2.0
filelock==3.8.2
Flask==2.1.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fonttools==4.38.0
future==0.18.2
fuzzysearch==0.7.3
gensim==4.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2020.1.16
httpretty==1.1.4
huggingface-hub==0.11.1
hypercorn==0.14.3
hyperframe==6.0.1
hyperopt==0.2.7
idna==3.3
ijson==2.6.1
importlib-metadata==4.8.2
importlib-resources==5.4.0
imutils==0.5.4
ipython==8.7.0
isodate==0.6.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
lml==0.1.0
lxml==4.6.3
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
msal==1.20.0
murmurhash==1.0.6
networkx==2.8.8
nltk==3.7
ntlm-auth==1.5.0
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
parse-email @ file:///C:/Users/bhumika.mahajan/Documents/Daimler/daimler_parse_email_application_win32/daimler_parse_email_application_win32/dist/parse_email-1.0.0_build_0.0.1-py3-none-any.whl
parso==0.8.3
pathlib==1.0.1
pathy==0.10.1
pbr==5.11.0
pdfkit==1.0.0
pep517==0.12.0
pickleshare==0.7.5
Pillow==9.1.0
pkg-about==1.0.3
plac==1.1.3
platformdirs==2.6.0
preshed==3.0.6
priority==2.0.0
prompt-toolkit==3.0.36
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.4.8
pyc-wheel==1.2.7
pycountry==22.3.5
pycparser==2.20
pydantic==1.8.2
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
Pygments==2.10.0
PyJWT==2.6.0
pymongo==3.12.0
pymsteams==0.2.2
pyparsing==3.0.6
PyPDF2==2.11.2
pysnow==0.7.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-http-client==3.3.7
python-magic==0.4.25
python-magic-bin==0.4.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==302
PyYAML==6.0
redis==4.4.0
regex==2022.10.31
requests==2.27.1
requests-file==1.5.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
rsa==4.7.2
sacremoses==0.0.53
schedule==1.1.0
scikit-learn==1.2.0
scikit-multilearn==0.2.0
scipy==1.9.3
seaborn==0.12.1
sendgrid==6.9.7
sentence-transformers==2.2.2
sentencepiece==0.1.96
servicenow==2.1.0
SharePlum==0.5.1
six==1.16.0
sklearn==0.0.post1
smart-open==6.2.0
SOAPpy==0.12.22
soupsieve==2.3.2.post1
spacy==3.2.1
spacy-legacy==3.0.10
spacy-loggers==1.0.4
srsly==2.4.5
stack-data==0.6.2
starkbank-ecdsa==2.2.0
stdiomask==0.0.6
sumy==0.11.0
textblob==0.17.1
texttable==1.6.7
thinc==8.0.13
threadpoolctl==3.1.0
tokenizers==0.13.2
toml==0.10.2
tomli==1.2.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
traitlets==5.7.0
transformers==4.17.0
twilio==7.15.4
typer==0.4.2
typing_extensions==4.4.0
tzdata==2021.2.post0
tzlocal==4.2
urllib3==1.25.11
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==2.2.2
wincertstore==0.2
wordcloud==1.8.2.2
wsproto==1.2.0
wstools==0.4.8
xgboost==1.5.2
xlrd==2.0.1
xmltodict==0.13.0
zeep==4.1.0
zipp==3.6.0
